requirements.txt
cassandra-driver==3.28.0
click==8.1.3
colorama==0.4.6
geomet==0.2.1.post1
kafka-python==2.0.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
six==1.16.0
